refactor(build/stapel): gate embedded stapel behind embedstapel build tag
Bare go build/vet/gopls and task test:unit previously required
running stapel:embed first, since pkg/stapel/embedded.go used a
top-level //go:embed on files that only exist after a network pull.
This broke any workflow outside the Taskfile build/lint/format
pipeline (IDE tooling, ad-hoc go test, CI jobs invoking go directly)
with a hard "pattern ... no matching files" compile error, and left
task test:unit undocumented as depending on stapel:embed despite
AGENTS.md mandating it as the canonical test entrypoint.

Split embedded image handling into three files: embedded_util.go
(platform normalization and checksum verification, tag-independent),
embedded.go (//go:build embedstapel, real go:embed directives and
docker load), and embedded_stub.go (//go:build !embedstapel, always
reports no embedded image so acquireImage falls back to the existing
docker pull path). Taskfile.dist.yaml passes the embedstapel tag only
through buildGoTags/buildCgoTags, applied to the actual build/dist/lint
targets; test:unit and bare go tooling keep using the untagged
goTags/cgoTags and no longer need embed artifacts on disk.

This also removes the need for an opt-out flag for slim/offline
builds (the tag itself is the opt-out) and closes the gap where a
release target could silently ship a pull-only binary if it forgot
to embed artifacts, since the tag decision now lives in a single
place (buildGoTags/buildCgoTags) instead of scattered per-target
tags.
